Added Slideshow State Machine
Added state machine to track file names for audio and image files,
length of audio files, transition types, and transition times. This can
be tied to the event handlers in the GUI, in addition I plan to add
seperate classes for saving the state data and parsing it. This way the
state machine can be functional on both the creator and presenter
applications.